working_samples/rag-pipeline/requirements.txt
torch==2.1.2
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
transformers==4.36.2
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.11
faiss-cpu==1.7.4
beautifulsoup4==4.12.3
unstructured==0.12.4
jq==1.6.0
black==24.2.0
# analysis-related
openpyxl
pandas
jupyterlab
numpy
matplotlib
word2number
python-dateutil
nltk
